HSEARCH-1765 moved DeleteByQuerySupport to impl
