+++ engine/src/main/java/org/hibernate/search/backend/impl/DeleteByQuerySupport.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend;
+package org.hibernate.search.backend.impl;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -23,6 +23,8 @@ import org.apache.lucene.search.BooleanClause.Occur;
 import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.BytesRefBuilder;
 import org.apache.lucene.util.NumericUtils;
+import org.hibernate.search.backend.DeletionQuery;
+import org.hibernate.search.backend.SingularTermQuery;
 import org.hibernate.search.backend.SingularTermQuery.Type;
 import org.hibernate.search.util.impl.ScopedAnalyzer;
