+++ engine/src/main/java/org/hibernate/search/backend/impl/lucene/works/DeleteByQueryWorkDelegate.java
@@ -13,10 +13,10 @@ import org.apache.lucene.search.BooleanQuery;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.TermQuery;
 import org.hibernate.search.backend.DeleteByQueryLuceneWork;
-import org.hibernate.search.backend.DeleteByQuerySupport;
 import org.hibernate.search.backend.IndexingMonitor;
 import org.hibernate.search.backend.LuceneWork;
 import org.hibernate.search.backend.DeletionQuery;
+import org.hibernate.search.backend.impl.DeleteByQuerySupport;
 import org.hibernate.search.engine.ProjectionConstants;
 import org.hibernate.search.exception.SearchException;
 import org.hibernate.search.store.Workspace;
