+++ serialization/avro/src/main/java/org/hibernate/search/indexes/serialization/avro/impl/AvroDeserializer.java
@@ -20,7 +20,7 @@ import org.apache.avro.generic.GenericRecord;
 import org.apache.avro.io.Decoder;
 import org.apache.avro.io.DecoderFactory;
 import org.apache.avro.util.Utf8;
-import org.hibernate.search.backend.DeleteByQuerySupport;
+import org.hibernate.search.backend.impl.DeleteByQuerySupport;
 import org.hibernate.search.bridge.spi.ConversionContext;
 import org.hibernate.search.bridge.util.impl.ContextualExceptionBridgeHelper;
 import org.hibernate.search.indexes.serialization.spi.Deserializer;
