+++ serialization/avro/src/main/java/org/hibernate/search/indexes/serialization/avro/impl/AvroSerializer.java
@@ -33,9 +33,9 @@ import org.apache.lucene.analysis.tokenattributes.TypeAttribute;
 import org.apache.lucene.index.IndexableField;
 import org.apache.lucene.util.AttributeImpl;
 import org.apache.lucene.util.BytesRef;
-import org.hibernate.search.backend.DeleteByQuerySupport;
 import org.hibernate.search.backend.DeletionQuery;
 import org.hibernate.search.backend.LuceneWork;
+import org.hibernate.search.backend.impl.DeleteByQuerySupport;
 import org.hibernate.search.indexes.serialization.spi.LuceneFieldContext;
 import org.hibernate.search.indexes.serialization.spi.LuceneNumericFieldContext;
 import org.hibernate.search.indexes.serialization.spi.Serializer;
